requirements.txt
biliup>=0.2.11
ffmpeg-python>=0.2.0
requests>=2.28.0
pytest>=7.0.0
